feat: add structure and start for training with crewai
Definitely requires some more polish, but the structure here should support everything. Next up is getting the yaml generated from the crew class defs, updating the mock inputs to what we want to test with (can even be multiple versions, but let's start with 1 solid example), then running the training and seeing results.

A lot of this still feels like a black box but we'll know more once we get a few runs done.